requirements.txt
aiofiles==25.1.0
certifi==2025.11.12
charset-normalizer==3.4.4
coloredlogs==15.0.1
flatbuffers==25.9.23
html5tagger==1.3.0
httptools==0.7.1
humanfriendly==10.0
idna==3.11
mpmath==1.3.0
multidict==6.7.0
numpy==2.2.6
onnxruntime==1.19.2
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
pillow-heif==0.22.0
protobuf==6.33.0
qiniu==7.14.0
requests==2.32.5
sanic==25.3.0
sanic-routing==23.12.0
setuptools==80.9.0
sympy==1.14.0
tracerite==1.1.3
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.6.1
uvloop==0.22.1
websockets==15.0.1
wheel==0.45.1
